Fix JS
On pages without maps (or where maps JS is not loaded),
`google` object does not exist.
Moving the map options (specifically mapTypeId) inside
a function will make sure we don't use `google` object
unless we actually need it to build a map.

This unbreaks this script and fixes masonry (and possibly
others) that failed because of it.